requirements.txt
streamlit
pandas
plotly
folium
streamlit-folium
pillow
numpy
openpyxl
scikit-learn
